body_len needs to be a size_t instead of int
2003-04-29  Jeremy Katz  <katzj@redhat.com>

	* providers/imap/camel-imap-folder.c (parse_fetch_response):
	body_len needs to be a size_t instead of int

svn path=/trunk/; revision=21027

+++ camel/providers/imap/camel-imap-folder.c
@@ -2579,7 +2579,7 @@ parse_fetch_response (CamelImapFolder *imap_folder, char *response)
 	GData *data = NULL;
 	char *start, *part_spec = NULL, *body = NULL, *uid = NULL, *idate = NULL;
 	gboolean cache_header = TRUE, header = FALSE;
-	int body_len = 0;
+	size_t body_len = 0;
 	
 	if (*response != '(') {
 		long seq;
